fitnesstracker/requirements.txt
asgiref==3.4.1
Django==4.0.2
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Pillow==9.0.1
psycopg2-binary==2.9.2
PyJWT==2.3.0
python-dotenv==0.17.1
pytz==2021.3
sqlparse==0.4.2
django-multiupload==0.6.1


